Add the initial python bindings.
This initial implementation uses the ctypes library to make dynamic calls to the
shared library.

I don't think this is the most pythonic API to do this with, but it wraps most
of the stuff we need it to.

It's also named KinectCore because all it does is wrap the C library.  More
user-friendly functions could be made in a "Kinect" class that uses this
KinectCore class, in the future.

Signed-off-by: Kelvie Wong <kelvie@ieee.org>